+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Adds the `rules:` option to `/lfg` which lets users stipulate additional rules for their game.
+
 ## [v15.2.1](https://github.com/lexicalunit/spellbot/releases/tag/v15.2.1) - 2025-07-02
 
 ### Fixed